refactor(capi): extract UTF-8/edit-anchoring helpers to CAPI_edit.cpp (todo 2.7)
Moved verbatim: the UTF-8 boundary walkers (utf8CharLen/nextChar/prevChar/
findAfter/findBefore + separator tables), getRange (now capi::getRange,
shared with the Interface overrides via DASHER_LOCAL decl in
CAPI_internal.h), caret clamping + validated sequence length + UTF-16/
codepoint byte-offset conversion, ensure_realized_for_context, and the
exported offset family (dasher_get_offset, byte_offset_from_utf16/
codepoints, dasher_set_offset, dasher_seed_buffer).
notify_buffer_cleared became an inline in CAPI_internal.h (used by both
TUs; same event value via DASHER_EVENT_BUFFER_CLEAR).
CAPI.cpp: 1540 -> 1194 lines. Verified: build [100%] 0 errors, 45/45
ctest, freeze 110.
