+++ CMakeLists.txt
@@ -221,6 +221,7 @@ if(BUILD_CAPI)
         ${CMAKE_CURRENT_LIST_DIR}/src/CAPI_appearance.cpp
         ${CMAKE_CURRENT_LIST_DIR}/src/CAPI_locale.cpp
         ${CMAKE_CURRENT_LIST_DIR}/src/CAPI_params.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/src/CAPI_edit.cpp
     )
 
     target_include_directories(dasher PRIVATE
@@ -390,7 +391,8 @@ dasher_add_test(dasher_text_metrics_tests test_text_metrics.cpp)
             ${CMAKE_CURRENT_LIST_DIR}/src/CAPI.cpp
             ${CMAKE_CURRENT_LIST_DIR}/src/CAPI_appearance.cpp
             ${CMAKE_CURRENT_LIST_DIR}/src/CAPI_locale.cpp
-            ${CMAKE_CURRENT_LIST_DIR}/src/CAPI_params.cpp)
+            ${CMAKE_CURRENT_LIST_DIR}/src/CAPI_params.cpp
+            ${CMAKE_CURRENT_LIST_DIR}/src/CAPI_edit.cpp)
         target_include_directories(dasher_control_action_tests PRIVATE
             ${CMAKE_CURRENT_LIST_DIR}/src/
             ${CMAKE_CURRENT_LIST_DIR}/tests/
